[refactor] calenderProvider 호출 구조 변경
