+++ app/(tabs)/medicine/register/_layout.tsx
@@ -1,0 +1,29 @@
+import { colors } from "@/constants";
+import Ionicons from "@expo/vector-icons/Ionicons";
+import { Stack, router } from "expo-router";
+import { TouchableOpacity } from "react-native";
+
+export default function RegisterLayout() {
+  return (
+    <Stack
+      screenOptions={{
+        headerShown: false,
+        contentStyle: { backgroundColor: colors.WHITE },
+      }}
+    >
+      <Stack.Screen
+        name="registerIndex"
+        options={{
+          headerShown: true,
+          title: "약물 등록",
+          contentStyle: { marginHorizontal: 13 },
+          headerLeft: () => (
+            <TouchableOpacity onPress={() => router.back()}>
+              <Ionicons name="chevron-back" size={24} color={"black"} />
+            </TouchableOpacity>
+          ),
+        }}
+      />
+    </Stack>
+  );
+}
