+++ app/(tabs)/medicine/register/registerIndex.tsx
